+++ admin/requires/menu-config.php
@@ -9,7 +9,7 @@
     </a>
     <ul class="nav nav-treeview">
         <li class="nav-item">
-        <a href="#" class="nav-link">
+        <a href="?config" class="nav-link">
             <i class="far fa-circle nav-icon"></i>
             <p>Dados da Página</p>
         </a>