+++ requires/about.php
@@ -15,7 +15,7 @@
                 <div class="body-cont mb-5 wow fadeInUp" data-wow-delay="0.6s">
                     <h3>Se te faz feliz e te dá bem-estar, você merece mais!  </h3>
                     <div class="bord-bottom"></div>
-                    <p>Profissionais que estão a anos no mercado da beleza, transformando a vida de muitas mulheres que procuram dar um Up no visual e na autoestima, a Bem Me Quero aguarda a tua presença, venha ser feliz também!</p>
+                    <p>Profissionais que estão a anos no mercado da beleza, transformando a vida de muitas mulheres que procuram dar um Up no visual e na autoestima, a <?php echo $Config->getName();?> aguarda a tua presença, venha ser feliz também!</p>
                     <a href="#contact" class="img-fluid js-scroll-trigger"><button class="btn btn-general btn-white">Contate-nos</button></a>
                 </div>
             </div>