+++ requires/perfectstyle.php
@@ -6,7 +6,7 @@
             <div class="col-md-6 my-auto">
                 <div class="perfect-style-desc text-center  wow fadeInLeft" data-wow-delay="0.3s">
                     <h3>Estilo Perfeito</h3>
-                    <p>A gente sabe que você adora passar horas no salão, mas com a vida tão corrida nem sempre é possível, não é mesmo? Pois, aqui na <i>Bem Me Quero</i> você pode ficar linda e economizar tempo! Agendamos cabeleireiro, manicure e pedicure no mesmo horário! Você linda da cabeça aos pés só na <i>Bem Me Quero</i>! </p>
+                    <p>A gente sabe que você adora passar horas no salão, mas com a vida tão corrida nem sempre é possível, não é mesmo? Pois, aqui na <i><?php echo $Config->getName();?></i> você pode ficar linda e economizar tempo! Agendamos cabeleireiro, manicure e pedicure no mesmo horário! Você linda da cabeça aos pés só na <i><?php echo $Config->getName();?></i>! </p>
                 </div>
             </div>
         </div>